After class - implemented UI functionality
Inserted the functions to take the option selected in the radio buttons and implemented the if statement to roll the dices. Implemented also an else to ask to select a dice. The result2 isVisible = true when display two results is clicked. Set result2 label to invisible again when display 1 result is selected.